Change colors, identificator label and pause/delete/edit algorithm
Changed colors on the UI for better user experience and differentiation
between items. The label "Identificator" from the CronTable has been
changed to "ID" for better spacing.

The algorithm for pausing, deleting, resuming and editing has been
adapted to be used when the cron job has no ID (which are the "default"
cron jobs comming from Linux)